requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
flask-cors
uvicorn
fastapi
mysql-connector-python